BigDecimal no controller

+++ app/src/main/java/com/example/cashswap/model/Moeda.kt
@@ -1,10 +1,12 @@
 package com.example.cashswap.model
 
+import java.math.BigDecimal
+
 data class Moeda(
     val codigo: String,
     val nome: String,
     val simbolo: String,
     val iconeResId: Int,
     val casasDecimais: Int,
-    var saldo: Double
+    var saldo: BigDecimal
 )
